Move ValidateLabels and ValidateAnnotations to utils package
