+++ pkg/function/utils.go
@@ -11,7 +11,6 @@ import (
 	"github.com/kanisterio/errkit"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/validation"
 	"k8s.io/client-go/kubernetes"
 
 	crv1alpha1 "github.com/kanisterio/kanister/pkg/apis/cr/v1alpha1"
@@ -21,6 +20,7 @@ import (
 	"github.com/kanisterio/kanister/pkg/log"
 	"github.com/kanisterio/kanister/pkg/param"
 	"github.com/kanisterio/kanister/pkg/secrets"
+	"github.com/kanisterio/kanister/pkg/utils"
 )
 
 const (
@@ -277,7 +277,7 @@ func ValidatePodLabelsAndAnnotations(funcName string, args map[string]any) error
 		return errkit.Wrap(err, "Kanister function validation failed, while getting pod labels from function args", "funcName", funcName)
 	}
 
-	if err = ValidateLabels(labels); err != nil {
+	if err = utils.ValidateLabels(labels); err != nil {
 		return errkit.Wrap(err, "Kanister function validation failed, while validating labels", "funcName", funcName)
 	}
 
@@ -285,7 +285,7 @@ func ValidatePodLabelsAndAnnotations(funcName string, args map[string]any) error
 	if err != nil {
 		return errkit.Wrap(err, "Kanister function validation failed, while getting pod annotations from function args", "funcName", funcName)
 	}
-	if err = ValidateAnnotations(annotations); err != nil {
+	if err = utils.ValidateAnnotations(annotations); err != nil {
 		return errkit.Wrap(err, "Kanister function validation failed, while validating annotations", "funcName", funcName)
 	}
 	return nil
@@ -331,29 +331,6 @@ func PodAnnotationsFromFunctionArgs(args map[string]any) (map[string]string, err
 		}
 	}
 	return nil, nil
-}
-
-func ValidateLabels(labels map[string]string) error {
-	for k, v := range labels {
-		if errs := validation.IsQualifiedName(k); len(errs) > 0 {
-			return errkit.New(fmt.Sprintf("label key '%s' failed validation. %s", k, errs))
-		}
-
-		if errs := validation.IsValidLabelValue(v); len(errs) > 0 {
-			return errkit.New(fmt.Sprintf("label value '%s' failed validation. %s", v, errs))
-		}
-	}
-	return nil
-}
-
-func ValidateAnnotations(annotations map[string]string) error {
-	for k := range annotations {
-		if errs := validation.IsQualifiedName(k); len(errs) > 0 {
-			return errkit.New(fmt.Sprintf("annotation key '%s' failed validation. %s", k, errs))
-		}
-	}
-	// annotation values don't actually have a strict format
-	return nil
 }
 
 type ActionSetAnnotations map[string]string
